Fixes padding in IFFT and IRFFT in sequence_layers.jax.
The padding/truncation used in FFT/IFFT and RFFT/IRFFT was always done prior to the transformation. This precludes the forward and backward transformations to be inverses of each other (when padding is used).

This CL reverses the order of padding/trunction for the backward transformations.
In this direction, it is not possible to always infer correctly the desired output length which must be provided as an extra parameter different from the fft length.

The tests were modified to reflect the corrected implementation's behavior.
New tests verifying the invertibility of the transforms were also added.

PiperOrigin-RevId: 736778995