Adds a SEMICAUSAL_FULL padding mode for sequence_layers.jax STFT/iSTFT that allows for perfect reconstruction.
Current padding available in sequence layer will either miss some initial frames of the STFT (e.g., reverse_causal) or introduce a delay between the input and output and lose trailing frames (e.g., causal).

Padding 'semicausal_full' makes sure that there is sufficient padding on the left and right of the sequence so that the input sequence can be reconstructed perfectly by InverseSTFT and that the mask is valid wherever the input signal is valid.
The output of InverseSTFT may be longer in terms of mask and tensor, but the valid values introduced are zero (in the absence of processing bewteen STFT and InverseSTFT).

PiperOrigin-RevId: 751226876